remove the line for inseting extra unused todo
We forgot that we already have one todo in the list from the previous
test so actually what we are clicking and testing in this particular
test is not what we added in thie line but the todo which was added
before.

So this line has no effect and therefore can be removed.